fixed Διαχειριστης 6) bug

+++ backend/admin5/deleteDiscount.js
@@ -5,17 +5,23 @@ const { ObjectId } = require('mongodb');
 const cache = require('../utils/cache.js');
 
 async function handleIndividualDiscountDeletion(req, res) {
+  if (req.session.user && req.session.user.isAdmin){
     try {
+      cache.flushAll();
       const discountId = req.query.discountId;
       const collection = await connectToDatabase("stock");
       const objectIdDiscountId = new ObjectId(discountId);
-      const result = await collection.updateOne({ _id: objectIdDiscountId }, { $set: { "discount": {}, "on_discount": false } });
-      cache.flushAll();
+      const result = await collection.updateOne({ _id: discountId }, { $set: { "discount": {}, "on_discount": false } });
       res.status(200).json(result);
+      // re-cache it
+      fetch(`http://localhost:3000/getDiscountedItems?shopId=${id}`);
     } catch (error) {
       console.error('Error deleting discount:', error);
       res.status(500).json({ error: 'Internal server error' });
     }
+  } else {
+    res.status(401).json({ error: 'Unauthorized' });
+  }
 }
 
 module.exports = handleIndividualDiscountDeletion;
